Use single instance of ConsoleOpSender
This is an alternative to figuring out how to make a ConsoleOpSender without logging the server version message. Technically, I think there should only be one ConsoleCommandSender, so it makes sense to only have one ConsoleOpSender.

If do I eventually figure out how to make a ConsoleOpSender without logging the server version message, I'm not yet sure if I would revert this commit or not.